add edition-lyon and edition-paris directories under models in publisher one
