+++ examples/EiffelClient.PublisherOne/Models/Edition-Paris/ActivityCanceled.cs
@@ -15,42 +15,30 @@
 using System;
 using EiffelEvents.Net.Events.Edition_Paris;
 
-namespace EiffelClient.PublisherOne.Models
+namespace EiffelClient.PublisherOne.Models.Edition_Paris
 {
-    public class EnvironmentDefined
+    public class ActivityCanceled
     {
-        public static EiffelEnvironmentDefinedEvent GetEvent()
+        public static EiffelActivityCanceledEvent GetEvent()
         {
-            return new EiffelEnvironmentDefinedEvent
+            return new EiffelActivityCanceledEvent
             {
-                Data = new()
+                Data = new ()
                 {
-                    Name = "env name",
-                    Image = "docker/image",
-                    Version = "2",
-                    Host = new()
-                    {
-                        Name = "name-1",
-                        User = "user-1"
-                    },
-                    CustomData = new()
-                    {
-                        { "key1", "test" },
-                        { "key2", new[] { 1, 2, 3 } }
-                    }
+                    Reason = "Made irrelevant by newly scheduled execution"
                 },
-                Meta = new()
+                Meta = new ()
                 {
                     Id = Guid.NewGuid().ToString(),
-                    Tags = new() { "docker env" },
-                    Security = new()
+                    Security = new ()
                     {
                         AuthorIdentity = "Flower"
                     }
                 },
-                Links = new()
+                Links = new ()
                 {
-                    PreviousVersion = new() { "cf056717-201b-43f6-9f2c-839b33b71baf" }
+                    ActivityExecution = "aaaaaaaa-bbbb-5ccc-8ddd-eeeeeeeeeee1",
+                    Cause = new () { "aaaaaaaa-bbbb-5ccc-8ddd-eeeeeeeeeee2" }
                 }
             };
         }
