+++ examples/EiffelClient.PublisherOne/Models/Edition-Paris/TestExecutionRecipeCollectionCreated.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using EiffelEvents.Net.Events.Edition_Paris;
 using EiffelEvents.Net.Events.Edition_Paris.Shared.Security;
 
-namespace EiffelClient.PublisherOne.Models
+namespace EiffelClient.PublisherOne.Models.Edition_Paris
 {
     public class TestExecutionRecipeCollectionCreated
     {
