+++ examples/EiffelClient.PublisherOne/Models/Edition-Paris/TestSuiteStarted.cs
@@ -18,7 +18,7 @@ using EiffelEvents.Net.Events.Edition_Paris;
 using EiffelEvents.Net.Events.Edition_Paris.Shared.Enums;
 using EiffelEvents.Net.Events.Edition_Paris.Shared.Security;
 
-namespace EiffelClient.PublisherOne.Models
+namespace EiffelClient.PublisherOne.Models.Edition_Paris
 {
     public class TestSuiteStarted
     {
